Restructing accounts app

+++ apps/accounts/models.py
@@ -6,7 +6,7 @@ from django.db import models
 
 
 class BaseModel(models.Model):
-    id = models.IntegerField(primary_key=True, editable=False, unique=True)
+    id = models.BigIntegerField(primary_key=True, editable=False, unique=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
